pyproject.toml
[project]
name = "conspiracy_bot"
version = "0.1.1"
description = "Telegram bot that returns random conspiracy theories"
authors = [{ name = "Laura Thompson", email = "laurthompson0@gmail.com" }]
dependencies = ["python-telegram-bot", "pyyaml", "requests"]
requires-python = ">=3.8"
license = { text = "_" }

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
